Builds project

+++ src/main/java/com/mauhwi/icantremembermybuilds/repo/ArmorRepository.java
@@ -3,5 +3,5 @@ package com.mauhwi.icantremembermybuilds.repo;
 import com.mauhwi.icantremembermybuilds.model.Esoarmor;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ArmorRepository extends CrudRepository<Esoarmor, Long> {
+public interface ArmorRepository extends CrudRepository<Esoarmor, Integer> {
 }
