+++ src/main/java/com/mauhwi/icantremembermybuilds/repo/WeaponsRepository.java
@@ -3,5 +3,5 @@ package com.mauhwi.icantremembermybuilds.repo;
 import com.mauhwi.icantremembermybuilds.model.Esoweapon;
 import org.springframework.data.repository.CrudRepository;
 
-public interface WeaponsRepository extends CrudRepository<Esoweapon, Long> {
+public interface WeaponsRepository extends CrudRepository<Esoweapon, Integer> {
 }
